requirements.txt
langchain-community
python-dotenv
unstructured
pypdf
openai
chromadb
tiktoken
pdfminer.six 
pdfminer
uuid
fastapi
typing
importlib
Markdown
nltk
langchain
langchain_openai
python-multipart